Clear claimed appointments if none are available
Previously, if an appointment became available, then _all_ appointments at
a location were taken, then that same appointment became available again,
it would not send a second notification for that appointment.

However if an appointment became available, then was taken, but there was
still other active appointments, then that appointment became available again,
it would have sent a second notication.

This fixes the former case to match the latter, so anytime an acceptable
appointment becomes available which was previously taken, it will send a
notification.